requirements.txt
flask==3.0.0
flask-cors==4.0.0
cryptography==41.0.7
PyJWT==2.8.0
